requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pdfplumber==0.11.1
anthropic==0.28.0
chromadb==0.5.3
python-multipart==0.0.9
streamlit==1.35.0
pandas==2.2.2
python-dotenv==1.0.1
httpx==0.27.0
